OO - cleaned up file structure
