+++ _archive/pickles_and_hard_data/geoms.py
@@ -1,0 +1,13 @@
+rooms_wkt_list = ["POLYGON ((85887.4870641205488937 -25151.4419914101599716, 64629.2000917164332350 -25151.4419914101599716, 65757.5078179416013882 -28251.4419914100362803, 79287.4870641207526205 -28251.4419914100362803, 85887.4870641205488937 -28251.4419914100362803, 85887.4870641205488937 -25151.4419914101599716))",
+"POLYGON ((64416.3645372212558868 -25151.4419914101599716, 60087.4870641206362052 -25151.4419914101599716, 60087.4870641206362052 -28251.4419914100362803, 65544.6722634464240400 -28251.4419914100362803, 64416.3645372212558868 -25151.4419914101599716))",
+"POLYGON ((60087.4870641206362052 -28451.4419914100326423, 60087.4870641206362052 -32751.4419914099125890, 64487.4870641206289292 -32751.4419914099125890, 79187.4870641207526205 -32751.4419914099125890, 79187.4870641207526205 -28451.4419914100326423, 65687.4870641206362052 -28451.4419914100326423, 60087.4870641206362052 -28451.4419914100326423))",
+"POLYGON ((85887.4870641205488937 -32751.4419914099125890, 85887.4870641205488937 -28451.4419914100326423, 79387.4870641207380686 -28451.4419914100326423, 79387.4870641207380686 -32751.4419914099125890, 85887.4870641205488937 -32751.4419914099125890))",
+"POLYGON ((85887.4870641205488937 -36951.4419914102036273, 85887.4870641205488937 -32951.4419914099125890, 79287.4870641207526205 -32951.4419914099125890, 64587.4870641206362052 -32951.4419914099125890, 64587.4870641206362052 -34651.4419914099125890, 79387.4870641207380686 -34651.4419914099635207, 79387.4870641207380686 -36951.4419914102036273, 85887.4870641205488937 -36951.4419914102036273))",
+"POLYGON ((60087.4870641206362052 -36951.4419914102036273, 79187.4870641207526205 -36951.4419914102036273, 79187.4870641207526205 -34851.4419914099635207, 64387.4870641206289292 -34851.4419914099125890, 64387.4870641206289292 -32951.4419914099125890, 60087.4870641206362052 -32951.4419914099125890, 60087.4870641206362052 -36951.4419914102036273))"]
+
+doors_wkt_list = ["POLYGON ((79437.4870641207380686 -36406.4419914101163158, 79437.4870641207380686 -35396.4419914101163158, 79137.4870641207380686 -35396.4419914101163158, 79137.4870641207380686 -36406.4419914101163158, 79437.4870641207380686 -36406.4419914101163158))",
+"POLYGON ((64637.4870641206362052 -34306.4419914099344169, 64637.4870641206362052 -33296.4419914099344169, 64337.4870641206362052 -33296.4419914099344169, 64337.4870641206362052 -34306.4419914099344169, 64637.4870641206362052 -34306.4419914099344169))",
+"POLYGON ((61982.4870641205998254 -33001.4419914099125890, 62992.4870641205998254 -33001.4419914099125890, 62992.4870641205998254 -32701.4419914099125890, 61982.4870641205998254 -32701.4419914099125890, 61982.4870641205998254 -33001.4419914099125890))",
+"POLYGON ((79137.4870641207526205 -31346.4419914098834852, 79137.4870641207526205 -32356.4419914098834852, 79437.4870641207526205 -32356.4419914098834852, 79437.4870641207526205 -31346.4419914098834852, 79137.4870641207526205 -31346.4419914098834852))",
+"POLYGON ((64793.2278058871524991 -26333.0555322997970507, 65138.6681506460663513 -27282.1450792935502250, 65420.5759368818398798 -27179.5390362958569312, 65075.1355921229260275 -26230.4494893020892050, 64793.2278058871524991 -26333.0555322997970507))",
+"POLYGON ((74382.4870641207235167 -28501.4419914100362803, 75392.4870641207235167 -28501.4419914100362803, 75392.4870641207235167 -28201.4419914100362803, 74382.4870641207235167 -28201.4419914100362803, 74382.4870641207235167 -28501.4419914100362803))"]
