+++ utils/simplify_poly.py
